Moved LoadGame to PathsOfPowerApp
